modify sql g4 file to use UDF instead of UDTF/UDSF/UDAF in register sql

+++ udf_funcs/python_scripts/udtf_multiply.py
@@ -5,7 +5,7 @@ class UDFMultiply(UDTFinDF):
     def eval(self, data):
         data = data.drop(columns=['key'])
         name = ', '.join(list(data))
-        res = data.prod(axis=1).to_frame(name=f'{self.udf_name}({name})')
+        res = data.prod(axis=1).to_frame(name=f'{self.udf_name}({name})').astype(float)
         return res
 
     # def __init__(self):
